// src/app/models/userLocation.ts
import { NgModel } from "@angular/forms";

export class UserLocationModal {
  query: string;
  status: string;
  country: string;
  countryCode: string;
  region: string;
  regionName: string;
  city: string;
  zip: string;
  lat: string;
  lon: string;
  timezone: string;
  isp: string;
  org: string;
  as: string;
}